@@ -1,64 +1,37 @@
 {
-  "name": "vue-antd-pro",
-  "version": "3.0.0",
+  "name": "abs-admin-vue",
+  "version": "4.0.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
-    "lint": "vue-cli-service lint",
-    "build:preview": "vue-cli-service build --mode preview",
-    "lint:nofix": "vue-cli-service lint --no-fix"
+    "build": "vue-cli-service build"
   },
   "dependencies": {
-    "@ant-design-vue/pro-layout": "^1.0.8",
-    "@antv/data-set": "^0.10.2",
-    "ant-design-vue": "^1.7.3",
-    "axios": "^0.19.0",
-    "core-js": "^3.1.2",
-    "enquire.js": "^2.1.6",
-    "lodash.clonedeep": "^4.5.0",
-    "lodash.get": "^4.4.2",
-    "lodash.pick": "^4.4.0",
-    "md5": "^2.2.1",
-    "mockjs2": "1.0.8",
-    "moment": "^2.24.0",
-    "nprogress": "^0.2.0",
+    "@ant-design/icons-vue": "^7.0.1",
+    "ant-design-vue": "^4.2.6",
+    "axios": "^1.7.9",
+    "core-js": "^3.8.3",
+    "dayjs": "^1.11.13",
+    "md5": "^2.3.0",
     "store": "^2.0.12",
-    "viser-vue": "^2.4.6",
-    "vue": "^2.6.10",
-    "vue-clipboard2": "^0.2.1",
-    "vue-cropper": "0.4.9",
-    "vue-i18n": "^8.17.4",
-    "vue-quill-editor": "^3.0.6",
-    "vue-router": "^3.1.2",
-    "vue-svg-component-runtime": "^1.0.1",
-    "vuex": "^3.1.1",
-    "wangeditor": "^3.1.1"
+    "vue": "^3.5.13",
+    "vue-router": "^4.5.0",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@ant-design/colors": "^3.2.1",
-    "@vue/cli-plugin-babel": "^4.0.4",
-    "@vue/cli-plugin-eslint": "^4.0.4",
-    "@vue/cli-plugin-router": "^4.0.4",
-    "@vue/cli-plugin-unit-jest": "^4.0.4",
-    "@vue/cli-plugin-vuex": "^4.0.4",
-    "@vue/cli-service": "^4.0.4",
-    "@vue/eslint-config-standard": "^4.0.0",
-    "@vue/test-utils": "^1.0.0-beta.29",
-    "babel-eslint": "^10.0.1",
-    "babel-plugin-import": "^1.12.2",
-    "babel-plugin-transform-remove-console": "^6.9.4",
-    "eslint": "^5.16.0",
-    "eslint-plugin-html": "^5.0.0",
-    "eslint-plugin-vue": "^5.2.3",
-    "git-revision-webpack-plugin": "^3.0.6",
-    "less": "^3.0.4",
-    "less-loader": "^5.0.0",
-    "opencollective": "^1.0.3",
-    "opencollective-postinstall": "^2.0.2",
-    "vue-svg-icon-loader": "^2.1.1",
-    "vue-template-compiler": "^2.6.10",
-    "webpack-theme-color-replacer": "^1.3.26"
-  }
+    "@vue/cli-plugin-router": "~5.0.0",
+    "@vue/cli-plugin-vuex": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "less": "^4.2.0",
+    "less-loader": "^12.0.0",
+    "tailwindcss": "^3.4.17",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.49"
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not dead",
+    "not ie 11"
+  ]
 }
