@@ -1,42 +1,11 @@
-// with polyfills
-import 'core-js/stable'
-import 'regenerator-runtime/runtime'
-
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
-import store from './store/'
-import i18n from './locales'
-import { VueAxios } from './utils/request'
-import ProLayout, { PageHeaderWrapper } from '@ant-design-vue/pro-layout'
-import themePluginConfig from '../config/themePluginConfig'
+import Antd from 'ant-design-vue'
+import 'ant-design-vue/dist/reset.css'
+import './assets/global.less'
 
-// mock
-// WARNING: `mockjs` NOT SUPPORT `IE` PLEASE DO NOT USE IN `production` ENV.
-import './mock'
-
-import bootstrap from './core/bootstrap'
-import './core/lazy_use' // use lazy load components
-import './permission' // permission control
-import './utils/filter' // global filter
-import './assets/global.less' // global style
-
-Vue.config.productionTip = false
-
-// mount axios to `Vue.$http` and `this.$http`
-Vue.use(VueAxios)
-// use pro-layout components
-Vue.component('pro-layout', ProLayout)
-Vue.component('page-container', PageHeaderWrapper)
-Vue.component('page-header-wrapper', PageHeaderWrapper)
-
-window.umi_plugin_ant_themeVar = themePluginConfig.theme
-
-window.vm = new Vue({
-  router,
-  store,
-  i18n,
-  // init localstorage, vuex
-  created: bootstrap,
-  render: h => h(App)
-}).$mount('#app')
+const app = createApp(App)
+app.use(router)
+app.use(Antd)
+app.mount('#app')
