@@ -1,83 +1,45 @@
 import axios from 'axios'
-import store from '@/store'
-import storage from 'store'
-import notification from 'ant-design-vue/es/notification'
-import { VueAxios } from './axios'
-import { ACCESS_TOKEN } from '@/store/mutation-types'
-import { showMsg } from '@/utils/data'
-import {VUE} from '@/main'
+
 // 创建 axios 实例
 const request = axios.create({
-  // API 请求的默认前缀
   baseURL: process.env.VUE_APP_API_BASE_URL,
-  timeout: 6000 // 请求超时时间
+  timeout: 5000
 })
 
-// 异常拦截处理器
-const errorHandler = (error) => {
-  if (error.response) {
-    const data = error.response.data
-    // 从 localstorage 获取 token
-    const token = storage.get(ACCESS_TOKEN)
-    if (error.response.status === 403) {
-      notification.error({
-        message: 'Forbidden',
-        description: data.message
-      })
+// 请求拦截器
+request.interceptors.request.use(
+  config => {
+    config.headers['Content-Type'] = 'application/json'
+    const token = localStorage.getItem('access_token')
+    if (token) {
+      config.headers['Authorization'] = token
     }
-    if (error.response.status === 401 && !(data.result && data.result.isLogin)) {
-      notification.error({
-        message: 'Unauthorized',
-        description: 'Authorization verification failed'
-      })
-      if (token) {
-        store.dispatch('Logout').then(() => {
-          setTimeout(() => {
-            window.location.reload()
-          }, 1500)
-        })
-      }
+    return config
+  },
+  error => {
+    console.error('Request Error:', error)
+    return Promise.reject(error)
+  }
+)
+
+// 响应拦截器
+request.interceptors.response.use(
+  response => {
+    const res = response.data
+    if (response.status !== 200) {
+      console.error('Error Response:', response.status)
+      return Promise.reject(new Error('http code=' + response.status || 'Error'))
     }
-  }
-  return Promise.reject(error)
-}
-
-// request interceptor
-request.interceptors.request.use(config => {
-  const token = storage.get(ACCESS_TOKEN)
-  // 如果 token 存在
-  // 让每个请求携带自定义 token 请根据实际情况自行修改
-  if (token) {
-    config.headers[ACCESS_TOKEN] = token
-  }
-  return config
-}, errorHandler)
-
-// response interceptor
-request.interceptors.response.use((response) => {
-  try{
-    if (response.headers['content-type'].indexOf("json") !== -1){
-      if (response.data.code !== 'SUCSESS'){
-        showMsg(window.vm,response.data);
-      }
-    }else{
-      return Promise.reject({response:response})
+    // 检查业务 code
+    if (res.code !== undefined && res.code !== '0') {
+      return Promise.reject({ response: { data: res } })
     }
-  }catch (e){
+    return res
+  },
+  error => {
+    console.error('Response Error:', error)
+    return Promise.reject(error)
   }
-  return response.data
-}, errorHandler)
-
-const installer = {
-  vm: {},
-  install (Vue) {
-    Vue.use(VueAxios, request)
-  }
-}
+)
 
 export default request
-
-export {
-  installer as VueAxios,
-  request as axios
-}
