@@ -1,325 +1,268 @@
 <template>
-  <div class="main">
-    <a-form
-      id="formLogin"
-      class="user-layout-login"
-      ref="formLogin"
-      :form="form"
-      @submit="handleSubmit"
-    >
-      <a-tabs
-        :activeKey="customActiveKey"
-        :tabBarStyle="{ textAlign: 'center', borderBottom: 'unset' }"
-        @change="handleTabClick"
+  <div class="min-h-screen flex items-center justify-center relative overflow-hidden bg-gradient-to-br from-[#667eea] to-[#764ba2]">
+    <!-- 背景装饰 -->
+    <div class="absolute inset-0 overflow-hidden">
+      <div class="absolute w-[300px] h-[300px] rounded-full bg-white/10 top-[-100px] left-[-100px] animate-float"></div>
+      <div class="absolute w-[200px] h-[200px] rounded-full bg-white/10 top-1/2 right-[-50px] animate-float-delay-1"></div>
+      <div class="absolute w-[150px] h-[150px] rounded-full bg-white/10 bottom-[-50px] left-[20%] animate-float-delay-2"></div>
+    </div>
+
+    <!-- 登录卡片 -->
+    <div class="relative w-full max-w-[420px] mx-6 bg-white/95 backdrop-blur-md rounded-2xl p-10 shadow-2xl animate-slide-up">
+      <!-- Logo 和标题 -->
+      <div class="text-center mb-10">
+        <img src="@/assets/logo.png" alt="Logo" class="w-16 h-16 mx-auto mb-4" />
+        <h1 class="text-3xl font-bold text-gray-900 mb-2">AbsAdmin</h1>
+        <p class="text-sm text-gray-500">后台管理系统</p>
+      </div>
+
+      <!-- 登录表单 -->
+      <a-form
+        :model="formState"
+        :rules="rules"
+        @finish="handleSubmit"
       >
-        <a-tab-pane key="tab1" :tab="$t('user.login.tab-login-credentials')">
-          <a-alert v-if="isLoginError" type="error" showIcon style="margin-bottom: 24px;" :message="$t('user.login.message-invalid-credentials')" />
-          <a-form-item>
-            <a-input
-              size="large"
-              type="text"
-              :placeholder="$t('user.login.username.placeholder')"
-              :maxLength="11"
-              v-decorator="[
-                'account',
-                {rules: [{ required: true, message: $t('user.userName.required') }, { validator: handleUsernameOrEmail }], validateTrigger: 'change'}
-              ]"
-            >
-              <a-icon slot="prefix" type="user" :style="{ color: 'rgba(0,0,0,.25)' }"/>
-            </a-input>
-          </a-form-item>
+        <a-alert
+          v-if="isLoginError"
+          type="error"
+          show-icon
+          message="账号或密码错误"
+          class="mb-6"
+        />
 
-          <a-form-item>
-            <a-input-password
-              size="large"
-              :placeholder="$t('user.login.password.placeholder')"
-              v-decorator="[
-                'password',
-                {rules: [{ required: true, message: $t('user.password.required') }], validateTrigger: 'blur'}
-              ]"
-            >
-              <a-icon slot="prefix" type="lock" :style="{ color: 'rgba(0,0,0,.25)' }"/>
-            </a-input-password>
-          </a-form-item>
-        </a-tab-pane>
-        <a-tab-pane key="tab2" :tab="$t('user.login.tab-login-mobile')">
-          <a-form-item>
-            <a-input size="large" type="text" :placeholder="$t('user.login.mobile.placeholder')" v-decorator="['mobile', {rules: [{ required: true, pattern: /^1[34578]\d{9}$/, message: $t('user.login.mobile.placeholder') }], validateTrigger: 'change'}]">
-              <a-icon slot="prefix" type="mobile" :style="{ color: 'rgba(0,0,0,.25)' }"/>
-            </a-input>
-          </a-form-item>
+        <a-form-item name="account">
+          <a-input
+            size="large"
+            v-model:value="formState.account"
+            placeholder="请输入账号"
+            :maxLength="11"
+            @change="handleAccountChange"
+          >
+            <template #prefix>
+              <UserOutlined class="text-gray-400 text-base" />
+            </template>
+          </a-input>
+        </a-form-item>
 
-          <a-row :gutter="16">
-            <a-col class="gutter-row" :span="16">
-              <a-form-item>
-                <a-input size="large" type="text" :placeholder="$t('user.login.mobile.verification-code.placeholder')" v-decorator="['captcha', {rules: [{ required: true, message: $t('user.verification-code.required') }], validateTrigger: 'blur'}]">
-                  <a-icon slot="prefix" type="mail" :style="{ color: 'rgba(0,0,0,.25)' }"/>
-                </a-input>
-              </a-form-item>
+        <a-form-item name="password">
+          <a-input-password
+            size="large"
+            v-model:value="formState.password"
+            placeholder="请输入密码"
+          >
+            <template #prefix>
+              <LockOutlined class="text-gray-400 text-base" />
+            </template>
+          </a-input-password>
+        </a-form-item>
+
+        <a-form-item name="vcode">
+          <a-row :gutter="12">
+            <a-col :span="14">
+              <a-input
+                size="large"
+                v-model:value="formState.vcode"
+                placeholder="请输入验证码"
+              >
+                <template #prefix>
+                  <SafetyOutlined class="text-gray-400 text-base" />
+                </template>
+              </a-input>
             </a-col>
-            <a-col class="gutter-row" :span="8">
-              <a-button
-                class="getCaptcha"
-                tabindex="-1"
-                :disabled="state.smsSendBtn"
-                @click.stop.prevent="getCaptcha"
-                v-text="!state.smsSendBtn && $t('user.register.get-verification-code') || (state.time+' s')"
-              ></a-button>
+            <a-col :span="10">
+              <div
+                v-if="captchaUrl"
+                class="w-full h-10 border border-gray-200 rounded-lg overflow-hidden cursor-pointer hover:border-[#667eea] hover:shadow-[0_0_0_2px_rgba(102,126,234,0.1)] transition-all bg-gray-50"
+                @click="refreshCaptcha"
+                title="点击刷新验证码"
+              >
+                <img :src="captchaUrl" class="w-full h-full object-cover" alt="验证码" />
+              </div>
             </a-col>
           </a-row>
-        </a-tab-pane>
-      </a-tabs>
+        </a-form-item>
 
-      <a-form-item>
-        <a-checkbox v-decorator="['rememberMe', { valuePropName: 'checked' }]">{{ $t('user.login.remember-me') }}</a-checkbox>
-        <router-link
-          :to="{ name: 'recover', params: { user: 'aaa'} }"
-          class="forge-password"
-          style="float: right;"
-        >{{ $t('user.login.forgot-password') }}</router-link>
-      </a-form-item>
+        <a-form-item class="mb-0">
+          <a-button
+            size="large"
+            type="primary"
+            html-type="submit"
+            class="w-full h-12 text-base font-semibold rounded-lg"
+            :loading="state.loginBtn"
+          >
+            {{ state.loginBtn ? '登录中...' : '登录' }}
+          </a-button>
+        </a-form-item>
+      </a-form>
 
-      <a-form-item style="margin-top:24px">
-        <a-button
-          size="large"
-          type="primary"
-          htmlType="submit"
-          class="login-button"
-          :loading="state.loginBtn"
-          :disabled="state.loginBtn"
-        >{{ $t('user.login.login') }}</a-button>
-      </a-form-item>
-
-      <div class="user-login-other">
-        <span>{{ $t('user.login.sign-in-with') }}</span>
-        <a>
-          <a-icon class="item-icon" type="alipay-circle"></a-icon>
-        </a>
-        <a>
-          <a-icon class="item-icon" type="taobao-circle"></a-icon>
-        </a>
-        <a>
-          <a-icon class="item-icon" type="weibo-circle"></a-icon>
-        </a>
-        <router-link class="register" :to="{ name: 'register' }">{{ $t('user.login.signup') }}</router-link>
+      <!-- 底部版权 -->
+      <div class="mt-8 text-center">
+        <p class="text-xs text-gray-500">© 2024 AbsAdmin. All rights reserved.</p>
       </div>
-    </a-form>
-
-    <two-step-captcha
-      v-if="requiredTwoStepCaptcha"
-      :visible="stepCaptchaVisible"
-      @success="stepCaptchaSuccess"
-      @cancel="stepCaptchaCancel"
-    ></two-step-captcha>
+    </div>
   </div>
 </template>
 
-<script>
+<script setup>
+import { ref, reactive, computed } from 'vue'
+import { useRouter } from 'vue-router'
+import { message, notification } from 'ant-design-vue'
 import md5 from 'md5'
-import TwoStepCaptcha from '@/components/tools/TwoStepCaptcha'
-import { mapActions } from 'vuex'
-import { timeFix } from '@/utils/util'
-import { getSmsCaptcha } from '@/api/login'
+import { UserOutlined, LockOutlined, SafetyOutlined } from '@ant-design/icons-vue'
+import { login } from '@/api/login'
 
-export default {
-  components: {
-    TwoStepCaptcha
-  },
-  data () {
-    return {
-      customActiveKey: 'tab1',
-      loginBtn: false,
-      // login type: 0 email, 1 account, 2 telephone
-      loginType: 0,
-      isLoginError: false,
-      requiredTwoStepCaptcha: false,
-      stepCaptchaVisible: false,
-      form: this.$form.createForm(this),
-      state: {
-        time: 60,
-        loginBtn: false,
-        // login type: 0 email, 1 account, 2 telephone
-        loginType: 0,
-        smsSendBtn: false
-      }
+const router = useRouter()
+
+const isLoginError = ref(false)
+const captchaTimestamp = ref(0)
+
+const formState = reactive({
+  account: '',
+  password: '',
+  vcode: ''
+})
+
+const state = reactive({
+  loginBtn: false
+})
+
+const rules = {
+  account: [{ required: true, message: '请输入账号' }],
+  password: [{ required: true, message: '请输入密码', trigger: 'blur' }],
+  vcode: [{ required: true, message: '请输入验证码', trigger: 'blur' }]
+}
+
+const captchaUrl = computed(() => {
+  if (!formState.account) return ''
+  return `/admin/captcha?account=${formState.account}&t=${captchaTimestamp.value}`
+})
+
+function handleAccountChange() {
+  refreshCaptcha()
+}
+
+function refreshCaptcha() {
+  captchaTimestamp.value = Date.now()
+}
+
+async function handleSubmit(values) {
+  state.loginBtn = true
+  try {
+    const loginParams = {
+      account: values.account,
+      password: md5(values.password),
+      vcode: values.vcode || ''
     }
-  },
-  created () {
- 
-  },
-  methods: {
-    ...mapActions(['Login', 'Logout']),
-    // handler
-    handleUsernameOrEmail (rule, value, callback) {
-      const { state } = this
-      const regex = /^([a-zA-Z0-9_-])+@([a-zA-Z0-9_-])+((\.[a-zA-Z0-9_-]{2,3}){1,2})$/
-      if (regex.test(value)) {
-        state.loginType = 0
-      } else {
-        state.loginType = 1
-      }
-      callback()
-    },
-    handleTabClick (key) {
-      this.customActiveKey = key
-      // this.form.resetFields()
-    },
-    handleSubmit (e) {
-      e.preventDefault()
-      const {
-        form: { validateFields },
-        state,
-        customActiveKey,
-        Login
-      } = this
 
-      state.loginBtn = true
-
-      const validateFieldsKey = customActiveKey === 'tab1' ? ['account', 'password'] : ['mobile', 'captcha']
-
-      validateFields(validateFieldsKey, { force: true }, (err, values) => {
-        if (!err) {
-          console.log('login form', values)
-          const loginParams = { ...values }
-          delete loginParams.account
-          loginParams[!state.loginType ? 'email' : 'account'] = values.account
-          loginParams.password = md5(values.password)
-          loginParams.vcode = ""
-          Login(loginParams)
-            .then((res) => this.loginSuccess(res))
-            .catch(err => this.requestFailed(err))
-            .finally(() => {
-              state.loginBtn = false
-            })
-        } else {
-          setTimeout(() => {
-            state.loginBtn = false
-          }, 600)
-        }
-      })
-    },
-    getCaptcha (e) {
-      e.preventDefault()
-      const { form: { validateFields }, state } = this
-
-      validateFields(['mobile'], { force: true }, (err, values) => {
-        if (!err) {
-          state.smsSendBtn = true
-
-          const interval = window.setInterval(() => {
-            if (state.time-- <= 0) {
-              state.time = 60
-              state.smsSendBtn = false
-              window.clearInterval(interval)
-            }
-          }, 1000)
-
-          const hide = this.$message.loading('验证码发送中..', 0)
-          getSmsCaptcha({ mobile: values.mobile }).then(res => {
-            setTimeout(hide, 2500)
-            this.$notification['success']({
-              message: '提示',
-              description: '验证码获取成功，您的验证码为：' + res.result.captcha,
-              duration: 8
-            })
-          }).catch(err => {
-            setTimeout(hide, 1)
-            clearInterval(interval)
-            state.time = 60
-            state.smsSendBtn = false
-            this.requestFailed(err)
-          })
-        }
-      })
-    },
-    stepCaptchaSuccess () {
-      this.loginSuccess()
-    },
-    stepCaptchaCancel () {
-      this.Logout().then(() => {
-        this.loginBtn = false
-        this.stepCaptchaVisible = false
-      })
-    },
-    loginSuccess (res) {
-      console.log(res)
-      // check res.homePage define, set $router.push name res.homePage
-      // Why not enter onComplete
-      /*
-      this.$router.push({ name: 'analysis' }, () => {
-        console.log('onComplete')
-        this.$notification.success({
-          message: '欢迎',
-          description: `${timeFix()}，欢迎回来`
-        })
-      })
-      */
-      this.$router.push({ path: '/' })
-      // 延迟 1 秒显示欢迎信息
-      setTimeout(() => {
-        this.$notification.success({
-          message: '欢迎',
-          description: `${timeFix()}，欢迎回来`
-        })
-      }, 1000)
-      this.isLoginError = false
-    },
-    requestFailed (err) {
-      this.isLoginError = true
-      this.$notification['error']({
-        message: '错误',
-        description: ((err.response || {}).data || {}).msg || '请求出现错误，请稍后再试',
-        duration: 4
-      })
+    const res = await login(loginParams)
+    if (res && res.data && res.data.access_token) {
+      localStorage.setItem('access_token', res.data.access_token)
     }
+    notification.success({
+      message: '欢迎',
+      description: '登录成功'
+    })
+    isLoginError.value = false
+    router.push('/')
+  } catch (err) {
+    isLoginError.value = true
+    notification.error({
+      message: '登录失败',
+      description: ((err.response || {}).data || {}).msg || '请求出现错误，请稍后再试',
+      duration: 4
+    })
+    refreshCaptcha()
+  } finally {
+    state.loginBtn = false
   }
 }
 </script>
 
-<style lang="less" scoped>
-.user-layout-login {
-  label {
-    font-size: 14px;
+<style scoped>
+/* 动画定义 */
+@keyframes float {
+  0%, 100% {
+    transform: translateY(0) rotate(0deg);
   }
-
-  .getCaptcha {
-    display: block;
-    width: 100%;
-    height: 40px;
+  50% {
+    transform: translateY(-20px) rotate(180deg);
   }
+}
 
-  .forge-password {
-    font-size: 14px;
+@keyframes slideUp {
+  from {
+    opacity: 0;
+    transform: translateY(30px);
   }
-
-  button.login-button {
-    padding: 0 15px;
-    font-size: 16px;
-    height: 40px;
-    width: 100%;
+  to {
+    opacity: 1;
+    transform: translateY(0);
   }
+}
 
-  .user-login-other {
-    text-align: left;
-    margin-top: 24px;
-    line-height: 22px;
+.animate-float {
+  animation: float 20s infinite ease-in-out;
+}
 
-    .item-icon {
-      font-size: 24px;
-      color: rgba(0, 0, 0, 0.2);
-      margin-left: 16px;
-      vertical-align: middle;
-      cursor: pointer;
-      transition: color 0.3s;
+.animate-float-delay-1 {
+  animation: float 20s infinite ease-in-out;
+  animation-delay: 5s;
+}
 
-      &:hover {
-        color: #1890ff;
-      }
-    }
+.animate-float-delay-2 {
+  animation: float 20s infinite ease-in-out;
+  animation-delay: 10s;
+}
 
-    .register {
-      float: right;
-    }
+.animate-slide-up {
+  animation: slideUp 0.5s ease-out;
+}
+
+/* Ant Design 组件样式覆盖 - 使用 :deep */
+:deep(.ant-input-affix-wrapper) {
+  border-radius: 8px;
+  border: 1px solid #e8e8e8;
+  padding: 12px 16px;
+  transition: all 0.3s;
+}
+
+:deep(.ant-input-affix-wrapper:hover) {
+  border-color: #667eea;
+}
+
+:deep(.ant-input-affix-wrapper-focused) {
+  border-color: #667eea;
+  box-shadow: 0 0 0 2px rgba(102, 126, 234, 0.1);
+}
+
+:deep(.ant-btn-primary) {
+  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+  border: none;
+  transition: all 0.3s;
+}
+
+:deep(.ant-btn-primary:hover:not(:disabled)) {
+  transform: translateY(-2px);
+  box-shadow: 0 8px 20px rgba(102, 126, 234, 0.4);
+}
+
+:deep(.ant-btn-primary:active:not(:disabled)) {
+  transform: translateY(0);
+}
+
+:deep(.ant-form-item) {
+  margin-bottom: 24px;
+}
+
+:deep(.ant-form-item-explain-error) {
+  font-size: 12px;
+}
+
+/* 响应式 */
+@media (max-width: 480px) {
+  .relative.w-full.max-w-\[420px\] {
+    margin: 16px;
+    padding: 32px 24px;
   }
 }
 </style>
