@@ -1,154 +1,12 @@
-const path = require('path')
-const webpack = require('webpack')
-const GitRevisionPlugin = require('git-revision-webpack-plugin')
-const GitRevision = new GitRevisionPlugin()
-const buildDate = JSON.stringify(new Date().toLocaleString())
-const createThemeColorReplacerPlugin = require('./config/plugin.config')
-
-function resolve (dir) {
-  return path.join(__dirname, dir)
-}
-
-// check Git
-function getGitHash () {
-  try {
-    return GitRevision.version()
-  } catch (e) {}
-  return 'unknown'
-}
-
-const isProd = process.env.NODE_ENV === 'production'
-const isDev = process.env.NODE_ENV === 'development'
-const isPreview = process.env.VUE_APP_PREVIEW === 'true'
-
-const assetsCDN = {
-  // webpack build externals
-  externals: {
-    vue: 'Vue',
-    'vue-router': 'VueRouter',
-    vuex: 'Vuex',
-    axios: 'axios'
-  },
-  css: [],
-  // https://unpkg.com/browse/vue@2.6.10/
-  js: [
-    'https://unpkg.com/vue@2.6.10/dist/vue.min.js',
-    'https://unpkg.com/vue-router@3.1.3/dist/vue-router.min.js',
-    'https://unpkg.com/vuex@3.1.1/dist/vuex.min.js',
-    'https://unpkg.com/axios@0.19.0/dist/axios.min.js'
-  ]
-}
-
-// vue.config.js
-const vueConfig = {
-  configureWebpack: config => {
-    const alias = {
-      '@$': resolve('src'),
-      '@public': resolve('public'),
-      '@assets': resolve('src/assets'),
-      '@components': resolve('src/components'),
-      '@config': resolve('src/config'),
-      '@common': resolve('src/common'),
-      '@mock': resolve('src/mock'),
-      '@views': resolve('src/views'),
-      '@store': resolve('src/store'),
-      '@utils': resolve('src/common/utils'),
-      '@apis': resolve('src/common/api')
-    }
-    Object.assign(config.resolve.alias, alias)
-
-    // 合并插件
-    const plugins = [...config.plugins]
-    plugins.push(
-      // Ignore all locale files of moment.js
-      new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
-      new webpack.DefinePlugin({
-        APP_VERSION: `"${require('./package.json').version}"`,
-        GIT_HASH: JSON.stringify(getGitHash()),
-        BUILD_DATE: buildDate
-      })
-    )
-    // preview.pro.loacg.com only do not use in your production;
-    if (isPreview) {
-      console.log('VUE_APP_PREVIEW', true)
-      // add `ThemeColorReplacer` plugin to webpack plugins
-      plugins.push(createThemeColorReplacerPlugin())
-    }
-    config.plugins = plugins
-    // if prod, add externals  //isProd ? assetsCDN.externals : {}
-    config.externals =  {}
-
-    if (isDev) config.devtool = 'source-map'
-  },
-
-  chainWebpack: (config) => {
-    const svgRule = config.module.rule('svg')
-    svgRule.uses.clear()
-    svgRule
-      .oneOf('inline')
-      .resourceQuery(/inline/)
-      .use('vue-svg-icon-loader')
-      .loader('vue-svg-icon-loader')
-      .end()
-      .end()
-      .oneOf('external')
-      .use('file-loader')
-      .loader('file-loader')
-      .options({
-        name: 'assets/[name].[hash:8].[ext]'
-      })
-
-    // if prod is on
-    // assets require on cdn
-    if (isProd) {
-      config.plugin('html').tap(args => {
-        //args[0].cdn = assetsCDN
-        return args
-      })
-    }
-  },
-
-  css: {
-    loaderOptions: {
-      less: {
-        modifyVars: {
-          // less vars，customize ant design theme
-
-          // 'primary-color': '#F5222D',
-          // 'link-color': '#F5222D',
-          'border-radius-base': '2px'
-        },
-        // DO NOT REMOVE THIS LINE
-        javascriptEnabled: true
-      }
-    }
-  },
-
+module.exports = {
+  publicPath: '/',
   devServer: {
-    // development server port 8000
-    port: 8001,
-    // If you want to turn on the proxy, please remove the mockjs /src/main.jsL11
-    // proxy: {
-    //   '/api': {
-    //     target: 'https://mock.ihx.me/mock/5baf3052f7da7e07e04a5116/antd-pro',
-    //     ws: false,
-    //     changeOrigin: true
-    //   }
-    // }
+    port: 8080,
     proxy: {
       '/admin': {
-        target: 'http://localhost:8000/',
-        ws: false,
+        target: 'http://localhost:8000',
         changeOrigin: true
       }
     }
-  },
-
-  // disable source map in production
-  productionSourceMap: false,
-  lintOnSave: false,
-  // babel-loader no-ignore node_modules/*
-  transpileDependencies: []
+  }
 }
-
-module.exports = vueConfig
